--- src/femini_reporting/src/start-http.ts
@@ -1,9 +1,0 @@
-import { server } from "./base";
-
-server.start({
-  transportType: "httpStream",
-  httpStream: {
-    port: 3002,
-    endpoint: "/stream",
-  },
-});
